56 clear Users_Class.jsx and add User functional component(Users.jsx) in UsersContainer.jsx for different flux(stream) datas
