+++ src/components/Users/Users.jsx
@@ -1,30 +1,36 @@
 import React from "react";
 import s from "./Users.module.css";
-import * as axios from "axios";
-import userPhoto from "./../../img/apa6.jpg";
+import userPhoto from "../../img/apa6.jpg";
+
+
+
 
 const Users = (props) => {
-  let	getUsers = () => {
-		if (props.users.length == 0) {
-			axios.get('https://social-network.samuraijs.com/api/1.0/users').then(response => {
-				props.setUsers(response.data.items)
-			})
-		}
+	let pageCount = Math.ceil(props.totalCount / props.pageSize);
+	let pages = [];
+	for (let i = 1; i <= pageCount; i++) {
+		pages.push(i);
 	}
-	return  <div className={s.users}>
-		<button onClick={getUsers}> Get Users</button>
+	return (
+		<div className={s.users}>
+			<div className={s.cursor_pointer}>
+				{pages.map(el => {
+					return	<span onClick={ (e) => props.onChangePages(el)} className={props.currentPage === el ? s.page_selected : null}> { el } </span>
+				})}
+			</div>
+			{/*<button className={s.getUsers}> Get Users</button>*/}
 			{props.users.map( (el) =>
 				<div className={s.users_items} key={el.id}>
 					<div className={s.users_div}>
+							<span className={s.users_span}>
+								{el.followed
+									? <button onClick={ () => {props.Follow(el.id)} }> Follow </button>
+									: <button onClick={ () => {props.unFollow(el.id)} }> Unfollow </button>
+								}
+							</span>
 						<span className={s.users_span}>
-							{el.followed
-								? <button onClick={ () => {props.Follow(el.id)} }> Follow </button>
-								: <button onClick={ () => {props.unFollow(el.id)} }> Unfollow </button>
-							}
-						</span>
-						<span className={s.users_span}>
-							<img className={s.users_span_img} src={el.photos.small != null ? el.photos.small : userPhoto} />
-						</span>
+								<img className={s.users_span_img} src={el.photos.small != null ? el.photos.small : userPhoto} />
+							</span>
 					</div>
 					<div className={s.users_div}>
 						<span className={s.users_span}>{el.name}</span>
@@ -39,6 +45,7 @@ const Users = (props) => {
 				</div>
 			)}
 		</div>
+	)
 }
 
-export default Users
+export default Users;
