+++ src/components/Users/UsersContainer.jsx
@@ -1,9 +1,40 @@
 import React from "react";
 import {connect} from "react-redux";
-import Users from "./Users_Class";
 import {changeCurrentPageAC, FollowAC, setTotalUsersCountAC, setUsersAC, unFollowAC} from "../../redux/Users_reducer";
+import * as axios from "axios";
+import Users from "./Users";
 
+class UsersContainer extends React.Component {
+	constructor(props) {
+		super(props);
+	}
+	componentDidMount() {
+		axios.get(`https://social-network.samuraijs.com/api/1.0/users&page=${this.props.currentPage}&count=${this.props.pageSize}`).then(response => {
+			this.props.setUsers(response.data.items)
+			this.props.setTotalUsersCount(response.data.totalCount)
 
+		})
+	}
+
+	onChangePages = (page) => {
+		this.props.changeCurrentPage(page);
+		axios.get(`https://social-network.samuraijs.com/api/1.0/users&page=${page}&count=${this.props.pageSize}`).then(response => {
+			this.props.setUsers(response.data.items)
+		})
+	}
+
+	render() {
+		return <Users totalCount={this.props.totalCount}
+									pageSize={this.props.pageSize}
+									currentPage={this.props.currentPage}
+									onChangePages={this.onChangePages}
+									users={this.props.users}
+									Follow={this.props.Follow}
+									unFollow={this.props.unFollow}
+
+		/>
+	}
+}
 
 let mapStateToProps = (state) => {
 	return {
@@ -34,6 +65,4 @@ let mapDispatchToProps = (dispatch) => {
 	}
 }
 
-let UsersContainer = connect(mapStateToProps,mapDispatchToProps)(Users);
-
-export default UsersContainer;
+export default connect(mapStateToProps,mapDispatchToProps)(UsersContainer);
